lld + (long long int)

+++ src/pdsi_ext.cpp
@@ -25,8 +25,8 @@ void pdsi::Rext_init(NumericVector& P, NumericVector& PE,
   calibrationEndYear = calib_e_yr;
 
   if(input_len != PE.length())
-    Rf_error("Length of input P (%d) is not equal to input PE (%ld).",
-             input_len, PE.length());
+    Rf_error("Length of input P (%d) is not equal to input PE (%lld).",
+             input_len, (long long int)PE.length());
 
   if(calibrationStartYear < startyear) {
     Rf_warning("Calibrating start year (%d) is earlier than start year (%d), "
